feat: add listening to hoodie account changes
